BJS2-65398:fix/fix comments

+++ src/main/java/faang/school/projectservice/service/MeetService.java
@@ -1,17 +1,19 @@
 package faang.school.projectservice.service;
 
 import faang.school.projectservice.dto.MeetDto;
+import faang.school.projectservice.exception.MeetNotFoundException;
+import faang.school.projectservice.exception.MeetWrongCreatorException;
 import faang.school.projectservice.mapper.MeetMapper;
 import faang.school.projectservice.model.Meet;
 import faang.school.projectservice.model.MeetStatus;
 import faang.school.projectservice.repository.MeetRepository;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -22,6 +24,12 @@ public class MeetService {
 
     public List<MeetDto> getAllMeets() {
         List<Meet> meets = meetRepository.findAll();
+
+        if (meets.isEmpty()) {
+            log.info("No meets found");
+            return List.of();
+        }
+
         return meets.stream()
                 .map(meetMapper::toDto)
                 .toList();
@@ -42,7 +50,7 @@ public class MeetService {
 
         return meets.stream()
                 .map(meetMapper::toDto)
-                .collect(Collectors.toList());
+                .toList();
     }
 
     public MeetDto getMeetById(long id) {
@@ -51,20 +59,19 @@ public class MeetService {
         return meetMapper.toDto(meet);
     }
 
-    public MeetDto createMeet(MeetDto meetDto) {
-        System.out.println("meetDto from service: " + meetDto);
+    public MeetDto createMeet(@NotNull MeetDto meetDto) {
         Meet meet = meetMapper.toEntity(meetDto);
         meet.setStatus(MeetStatus.PENDING);
         meet = meetRepository.save(meet);
+        log.info("Meet with id {} created", meet.getId());
         return meetMapper.toDto(meet);
     }
 
-    public MeetDto updateMeet(long id, MeetDto meetDto) {
+    public MeetDto updateMeet(long id, @NotNull MeetDto meetDto, long userId) {
         Meet meet = meetRepository.findById(id)
                 .orElseThrow(() -> meetNotFoundException(id));
         if (meet.getCreatorId() != meetDto.creatorId()) {
-            log.error("Meet with id {} is not created by user with id {}", id, meetDto.creatorId());
-            throw new IllegalArgumentException("Only creator can update the meet");
+            throw meetWrongCreatorException(id, userId);
         }
         meet.setTitle(meetDto.title());
         meet.setDescription(meetDto.description());
@@ -74,20 +81,32 @@ public class MeetService {
         return meetMapper.toDto(meet);
     }
 
-    public MeetDto cancelMeet(long id) {
+    public MeetDto cancelMeet(long id, long userId) {
         Meet meet = meetRepository.findById(id)
                 .orElseThrow(() -> meetNotFoundException(id));
+        if (meet.getCreatorId() != userId) {
+            throw meetWrongCreatorException(id, userId);
+        }
         meet.setStatus(MeetStatus.CANCELLED);
         meet = meetRepository.save(meet);
         return meetMapper.toDto(meet);
     }
 
-    public void deleteMeet(long id) {
+    public void deleteMeet(long id, long userId) {
+        Meet meet = meetRepository.findById(id).orElseThrow(() -> meetNotFoundException(id));
+        if (meet.getCreatorId() != userId) {
+            throw meetWrongCreatorException(id, userId);
+        }
         meetRepository.deleteById(id);
     }
 
-    private IllegalArgumentException meetNotFoundException(long id) {
+    private MeetNotFoundException meetNotFoundException(long id) {
         log.error("Meet with id {} not found", id);
-        return new IllegalArgumentException("Meet not found");
+        return new MeetNotFoundException(id);
+    }
+
+    private MeetWrongCreatorException meetWrongCreatorException(long id, long userId) {
+        log.error("Meet with id {} is not created by user with id {}", id, userId);
+        return new MeetWrongCreatorException("Only creator can update the meet");
     }
 }
